chore: apply migrations, configure venue policy, and backfill event data
Manual configuration for initial production setup:
- Fix FK type mismatch in CRM migration (contacts.id is bigint(20))
- Apply all 9 new tenant migrations (026-030)
- Generate CREDENTIAL_ENCRYPTION_KEY and encrypt 1 existing credential row
- Insert venue policy: upstairs 450 all-ages, downstairs 350 21+; COI required,
  25% deposit, doors 5pm, curfew 2am
- Backfill is_private=1 on all event_type='private_event' events (2 events)
- All 217 existing events set to deposit_status='not_required' for backward compat
- Add roadmap follow-up notes (POS, payroll, accounting, client portal, CRM email,
  incident workflow, mobile offline, payment links, Stripe Connect, key rotation)

+++ database/migrations/tenant/014_crm.sql
@@ -36,7 +36,7 @@ CREATE TABLE IF NOT EXISTS `client_profiles` (
   `notes`               TEXT          DEFAULT NULL,
   `consent_marketing`   TINYINT(1)    NOT NULL DEFAULT 0,
   `consent_date`        DATE          DEFAULT NULL,
-  `contact_id`          INT(11)       DEFAULT NULL,
+  `contact_id`          BIGINT(20)    DEFAULT NULL,
   `created_by_id`       INT(11)       DEFAULT NULL,
   `created_at`          TIMESTAMP     NOT NULL DEFAULT CURRENT_TIMESTAMP,
   `updated_at`          TIMESTAMP     NOT NULL DEFAULT CURRENT_TIMESTAMP ON UPDATE CURRENT_TIMESTAMP,
